card.ts file moved to util folder
