+++ src/app/diary-module/util/card.ts
